requirements.txt
keras==3.10.0
metrics==0.0.2
nltk==3.9.1
numpy==2.1.3
pandas==2.3.0
pillow==11.2.1
regex==2024.11.6
scikit-learn==1.7.0
streamlit==1.46.0
tensorflow==2.19.0